basically LL is a one side love story.
Assume a, b, c, d are references  of the objects created. a,b,c,d are linked so a knows the address of b but b doesn't knows the address of a and so on.. This a one sided love story where a is in love with b but b is in love with c and c is in love with d and d have a null love story.